v1/requirements.txt
numpy
pandas
flask
gunicorn
flask-cors
python-dotenv
authlib
psycopg2-binary
sqlalchemy>=2.0
bcrypt
cmake
opencv-python-headless==4.5.5.62
face-recognition==1.3.0
requests==2.31.0
Pillow==9.3.0
google-api-python-client==2.54.0
pytest==6.2.5
flask
gunicorn
msal
flask-cors
google.generativeai


